sending info serve-client partial
