+++ Assets/Scripts/Game/PathFinding/PathFinding.cs
@@ -1,6 +1,5 @@
 using System.Collections;
 using System.Collections.Generic;
-using UnityEditor.U2D.Path.GUIFramework;
 using UnityEngine;
 
 public class PathFinding : Singleton<PathFinding>
